# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ChefRole": {
        "enum": ["SOUS_CHEF", "HEAD_CHEF"],
        "type": "string",
        "x-shared-type": "ChefRoleValue"
      },
...
      "SystemRole": {
        "enum": ["ROOT_ADMIN"],
        "type": "string",
        "x-shared-type": "SystemRoleValue"
      },
      "User": {
        "additionalProperties": true,
        "properties": {
          "accountStatus": { "$ref": "#/components/schemas/UserAccountStatus" },
          "avatarUrl": { "nullable": true, "type": "string" },
          "chefRole": {
            "anyOf": [{ "$ref": "#/components/schemas/ChefRole" }],
            "nullable": true
          },
          "createdAt": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "name": { "type": "string" },
          "paymentRemindersEnabled": { "type": "boolean" },
          "phone": { "nullable": true, "type": "string" },
          "roles": { "items": { "type": "string" }, "type": "array" },
          "systemRole": {
            "anyOf": [{ "$ref": "#/components/schemas/SystemRole" }],
            "nullable": true
          },
          "username": { "type": "string" }
        },
        "required": [
          "id",
          "username",
          "name",
          "chefRole",
          "systemRole",
          "accountStatus",
          "roles"
        ],
        "type": "object"
      },
      "UserAccountStatus": {
        "enum": ["ACTIVE", "BLOCKED"],
        "type": "string",
        "x-shared-type": "UserAccountStatusValue"
      }
    },
    "securitySchemes": { "bearerAuth": { "scheme": "bearer", "type": "http" } }
  },
  "info": { "title": "FF RESTaurent API", "version": "1.1.0" },
  "openapi": "3.0.3",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "postAuthLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "identifier": { "minLength": 1, "type": "string" },
                  "password": { "minLength": 1, "type": "string" }
                },
                "required": ["identifier", "password"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "token": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/User" }
                  },
                  "required": ["token", "user"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "2XX": {
            "content": { "application/json": { "schema": {} } },
            "description": "Default Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": []
      }
    },
...
    "/auth/register": {
      "post": {
        "operationId": "postAuthRegister",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "inviteCode": { "minLength": 1, "type": "string" },
                  "name": { "minLength": 1, "type": "string" },
                  "password": { "minLength": 8, "type": "string" },
                  "phone": {
                    "anyOf": [
                      { "maxLength": 40, "type": "string" },
                      { "enum": ["null"], "nullable": true }
                    ]
                  },
                  "username": {
                    "maxLength": 30,
                    "minLength": 3,
                    "type": "string"
                  }
                },
                "required": ["name", "username", "password", "inviteCode"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "token": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/User" }
                  },
                  "required": ["token", "user"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "2XX": {
            "content": { "application/json": { "schema": {} } },
            "description": "Default Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": []
      }
    },
...
  }
}