# apps/api/src/routes/auth-routes.ts
import type { FastifyInstance } from 'fastify';
import bcrypt from 'bcryptjs';
import { parseVietnamMobilePhone } from '@ff-restaurent/shared';
import { prisma } from '../lib/prisma.js';
import { sanitizeUser } from '../lib/roles.js';
import { loginSchema, registerSchema } from '../schemas/index.js';
import { ensureDefaultCollections } from '../services/collection-service.js';
import { UserAccountStatus } from '@prisma/client';

/**
 * Authentication routes issue JWTs and return sanitized user profiles.
 */
export const registerAuthRoutes = (app: FastifyInstance) => {
  const authRateLimit = { max: 5, timeWindow: '1 minute' };

  app.post(
    '/auth/login',
    { config: { rateLimit: authRateLimit } },
    async (request, reply) => {
      const body = loginSchema.parse(request.body);
      let user = await prisma.user.findUnique({
        where: { username: body.identifier },
      });
      if (!user) {
        const parsedPhone = parseVietnamMobilePhone(body.identifier);
        if (parsedPhone.success && parsedPhone.phone) {
          user = await prisma.user.findUnique({
            where: { phone: parsedPhone.phone },
          });
        }
      }
      const passwordMatches =
        user && (await bcrypt.compare(body.password, user.passwordHash));
      if (
        !user ||
        !passwordMatches ||
        user.accountStatus !== UserAccountStatus.ACTIVE
      ) {
        return reply.code(401).send({
          code: 'INVALID_CREDENTIALS',
          message: 'Invalid credentials',
        });
      }
      await ensureDefaultCollections(user.id);
      request.log.info({ event: 'login_succeeded', userId: user.id });
      return {
        token: app.jwt.sign({ sub: user.id, ver: user.sessionVersion }),
        user: sanitizeUser(user),
      };
    },
  );

  app.post(
    '/auth/register',
    { config: { rateLimit: authRateLimit } },
    async (request, reply) => {
      const body = registerSchema.parse(request.body);
      if (body.inviteCode !== app.config.registrationInviteCode) {
        return reply.code(403).send({
          code: 'REGISTRATION_NOT_AUTHORIZED',
          message: 'Registration is not authorized',
        });
      }
      const existing = await prisma.user.findFirst({
        where: {
          OR: [
            { username: body.username },
            ...(body.phone ? [{ phone: body.phone }] : []),
          ],
        },
      });
      if (existing) {
        return reply.code(409).send({
          code: 'IDENTIFIER_TAKEN',
          message: 'Username or phone already taken',
        });
      }
      const user = await prisma.user.create({
        data: {
          name: body.name,
          username: body.username,
          phone: body.phone ?? null,
          passwordHash: await bcrypt.hash(body.password, 12),
        },
      });
      await ensureDefaultCollections(user.id);
      return reply.code(201).send({
        token: app.jwt.sign({ sub: user.id, ver: user.sessionVersion }),
        user: sanitizeUser(user),
      });
    },
  );
};